Update and fix components
* Install "style-loader" and add this module in "webpack.dev.config.js"
* Update index.tsx
* Remove "src/components/App" folder
* Create new App.tsx file in "src/components/"
* Create new component "Logo" in "src/components/shared/Logo"
* Create new component "Navigation" in "src/components/shared/Navbar/Navigation/"
* Create new component "Search" in "src/components/shared/Navbar/Search/"
* Create and connecte "module.css" in new components
* Trouble: application does not connect "module.css" files
