+++ src/components/shared/Logo/Logo.module.css
@@ -1,0 +1,4 @@
+.logo {
+	font-size: 22px;
+	text-decoration: none;
+}
