+++ src/components/shared/Navbar/Navbar.module.css
@@ -1,0 +1,3 @@
+.navbar {
+	background: #1E5A70;
+}
