+++ src/components/shared/Navbar/Search/Search.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+class Search extends React.Component {
+	render() {
+		return (
+			<form action="">
+				<input type="search" id="search" name="search"></input>
+			</form>
+		);
+	}
+}
+
+export default Search;
