+++ src/components/shared/Navbar/Search/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./Search";
