+++ webpack.dev.config.ts
@@ -46,6 +46,10 @@ const config: Configuration = {
           loader: "html-loader",
         },
       },
+      {
+        test: /\.css$/i,
+        use: ["style-loader", "css-loader"],
+      }
     ],
   },
   resolve: {
